clear and make component and split element
