--- src/routes/homeadmin/test/+page.svelte
@@ -1,4 +1,0 @@
-<h1>Hello form User</h1>
-<a href="/homeadmin">
-  <img src="https://media1.tenor.com/m/ue7Q8JmP_0MAAAAd/oiia-oiiaoiia.gif" alt="Admin GIF" style="width:200px;" />
-</a>
